Move types to a types folder
