--- src/itemType.ts
@@ -1,1 +1,0 @@
-export type ItemType = string;
