+++ src/types/characterClass.ts
