+++ src/types/items.ts
@@ -1,1 +1,3 @@
+export type ItemType = string;
+
 export type ItemEffect = string;
